+++ src/HtmlLabel/iOS/Renderer.cs
@@ -37,7 +37,7 @@ namespace LabelHtml.Forms.Plugin.iOS
 			}
 			catch (System.Exception ex)
 			{
-				System.Diagnostics.Debug.WriteLine(@$"ERROR: ${ex.Message}");
+				System.Diagnostics.Debug.WriteLine(@"            ERROR: ", ex.Message);
 			}
 		}
 
@@ -53,7 +53,7 @@ namespace LabelHtml.Forms.Plugin.iOS
 				}
 				catch (System.Exception ex)
 				{
-					System.Diagnostics.Debug.WriteLine(@$"ERROR: ${ex.Message}");
+					System.Diagnostics.Debug.WriteLine(@"            ERROR: ", ex.Message);
 				}
 			}
 		}